refactor: clean code and better encapsulation
